requirements.txt
asgiref==3.6.0
backports-zoneinfo==0.2.1;python_version<"3.9"
blis==0.7.9
build==0.10.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
cryptography==39.0.2
cymem==2.0.7
defusedxml==0.7.1
dj-rest-auth==2.2.3
django==4.1.7
django-allauth==0.50.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==21.1
django-storages==1.13.2
django-stubs==1.9.0
django-stubs-ext==0.7.0
djangorestframework==3.12.4
idna==3.4
inflect==6.0.2
jinja2==3.1.2
langcodes==3.3.0
markupsafe==2.1.2
murmurhash==1.0.9
mypy==1.0.1
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
preshed==3.0.8
pycparser==2.21
pydantic==1.10.5
pyjwt[crypto]==2.6.0
pyproject-hooks==1.0.0
pyspellchecker==0.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
thinc==8.1.9
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
typer==0.7.0
types-pytz==2022.7.1.2
types-pyyaml==6.0.12.8
typing-extensions==4.5.0
urllib3==1.26.14
wasabi==1.1.1
wheel==0.38.4
